EQF. jpynb.
The goal of this project is to build a predictive model that can
accurately classify individuals as either having or not having heart
disease. By analysing a comprehensive set of patient attributes,
including demographic, clinical, and lifestyle factors, the model aims
to identify patterns and features indicative of the presence or
absence of heart disease.